remove URI coercion in log file repair
If the source file is in the current directory, the path is coerced into the authority
part of the URI. This causes Paths/get to blow up, since the URI has no path segment.

We can avoid URIs entirely because we already have File handles for both the original
log file and the repaired log file.